+++ mac/Tests/CTVersionTests.m
@@ -31,7 +31,8 @@
     
     const uint32_t vhex = ct_versionhex(&v);
     
-    const uint32_t expected = 0 | v.major << 16 | v.minor << 8 | v.patch;
+    const uint32_t expected = (uint32_t)(0 | v.major << 16 | v.minor << 8
+                                         | v.patch);
     XCTAssertEqual(expected, vhex);
 }
